+++ gen/org/elixir_lang/psi/ElixirHeredoc.java
@@ -9,8 +9,9 @@ import com.ericsson.otp.erlang.OtpErlangObject;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.LiteralTextEscaper;
 import com.intellij.psi.PsiLanguageInjectionHost;
+import com.intellij.util.concurrency.annotations.RequiresReadLock;
 
-public interface ElixirHeredoc extends HeredocLiteral, Quote {
+public interface ElixirHeredoc extends HeredocLiteral, Interpolated, Quote {
 
   @NotNull
   List<ElixirHeredocLine> getHeredocLineList();
@@ -20,8 +21,10 @@ public interface ElixirHeredoc extends HeredocLiteral, Quote {
 
   @NotNull List<Integer> addEscapedCharacterCodePoints(@Nullable List<Integer> codePointList, @NotNull ASTNode child);
 
+  @RequiresReadLock
   @NotNull List<Integer> addEscapedEOL(@Nullable List<Integer> maybeCodePointList, @NotNull ASTNode child);
 
+  @RequiresReadLock
   @NotNull List<Integer> addEscapedTerminator(@Nullable List<Integer> maybeCodePointList, @NotNull ASTNode child);
 
   @NotNull List<Integer> addFragmentCodePoints(@Nullable List<Integer> codePointList, @NotNull ASTNode child);