+++ gen/org/elixir_lang/psi/ElixirLiteralSigilHeredoc.java
@@ -9,6 +9,7 @@ import com.ericsson.otp.erlang.OtpErlangObject;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.LiteralTextEscaper;
 import com.intellij.psi.PsiLanguageInjectionHost;
+import com.intellij.util.concurrency.annotations.RequiresReadLock;
 
 public interface ElixirLiteralSigilHeredoc extends Literal, SigilHeredocLiteral {
 
@@ -23,8 +24,10 @@ public interface ElixirLiteralSigilHeredoc extends Literal, SigilHeredocLiteral 
 
   @NotNull List<Integer> addEscapedCharacterCodePoints(@Nullable List<Integer> codePointList, @NotNull ASTNode child);
 
+  @RequiresReadLock
   @NotNull List<Integer> addEscapedEOL(@Nullable List<Integer> maybeCodePointList, @NotNull ASTNode child);
 
+  @RequiresReadLock
   @NotNull List<Integer> addEscapedTerminator(@Nullable List<Integer> maybeCodePointList, @NotNull ASTNode child);
 
   @NotNull List<Integer> addFragmentCodePoints(@Nullable List<Integer> codePointList, @NotNull ASTNode child);