+++ src/org/elixir_lang/psi/impl/ElixirPsiImplUtil.java
@@ -1801,6 +1801,7 @@ public class ElixirPsiImplUtil {
         return ParentImpl.addEscapedCharacterCodePoints(parent, codePointList, child);
     }
 
+    @RequiresReadLock
     @NotNull
     public static List<Integer> addEscapedEOL(@NotNull Parent parent,
                                               @Nullable List<Integer> maybeCodePointList,
@@ -1808,11 +1809,12 @@ public class ElixirPsiImplUtil {
         return ParentImpl.addEscapedEOL(parent, maybeCodePointList);
     }
 
+    @RequiresReadLock
     @NotNull
     public static List<Integer> addEscapedTerminator(@NotNull Parent parent,
                                                      @Nullable List<Integer> maybeCodePointList,
                                                      @NotNull ASTNode child) {
-        return ParentImpl.addEscapedTerminator(maybeCodePointList, child);
+        return ParentImpl.addEscapedTerminator(parent, maybeCodePointList, child);
     }
 
     @NotNull