+++ tests/org/elixir_lang/parser_definition/LiteralWordsHeredocParsingTestCase.java
@@ -1,5 +1,6 @@
 package org.elixir_lang.parser_definition;
 
+import org.elixir_lang.psi.quoting.QuotingDialect;
 /**
  * Created by kadie.enheduanna.inanna on 8/8/14.
  */
@@ -33,7 +34,7 @@ public class LiteralWordsHeredocParsingTestCase extends ParsingTestCase {
     }
 
     public void testWhitespaceEndPrefix() {
-        assertParsedAndQuotedCorrectly();
+        assertParsedAndQuotedCorrectlyFrom(QuotingDialect.V1_12);
     }
 
     @Override